+++ src/lib/google-analytics/index.ts
@@ -1,0 +1,2 @@
+export { GoogleAnalytics, usePageView } from "./gtag";
+export { runReport } from "./report";
