+++ src/services/article.ts
@@ -1,6 +1,6 @@
+import { runReport } from "@/lib/google-analytics";
 import type { TArticle, TRankedArticle, TTag } from "@/types";
 import { fetchArticles, fetchPickupArticles } from "@/utils/fetcher";
-import { runReport } from "@/utils/ga/report";
 
 /**
  * 走査対象のタグを重複なし2個ずつにグループ化
